move all plotting functions
